Move epsg constants to a single file
